implement book management functionality with CRUD operations and API integration
