+++ frontend/src/store/book.js
@@ -1,32 +1,38 @@
 import { create } from "zustand";
 
-export const useBookStore = create((set) => ({
-  // Books state
+const API_URL = "http://localhost:5000/api";
+
+export const useBookStore = create((set, get) => ({
   books: [],
   booksLoading: false,
   booksError: null,
-  setNewBook: (books) => set({ books }),
+  clusterId: null,
 
-  // Popular books state
-  popularBooks: [],
-  popularBooksLoading: false,
-  popularBooksError: null,
-  setPopularBooks: (popularBooks) => set({ popularBooks }),
+  setClusterId: (clusterId) => set({ clusterId }),
 
-  // Create book
   createBook: async (newBook) => {
-    const requiredFields = ['title', 'author', 'description', 'genre', 'status', 'image', 'reviews'];
-    const missingFields = requiredFields.filter(field => !newBook[field]);
-    
+    const requiredFields = [
+      "title",
+      "author",
+      "description",
+      "genre",
+      "status",
+      "image",
+      "clusterId",
+    ];
+    const missingFields = requiredFields.filter((field) => !newBook[field]);
+
     if (missingFields.length > 0) {
-      return { 
-        success: false, 
-        message: `Missing required fields: ${missingFields.join(', ')}` 
+      return {
+        success: false,
+        message: `Missing required fields: ${missingFields.join(", ")}`,
       };
     }
 
     try {
-      const res = await fetch("/api/books", {
+      console.log("Sending book data with clusterId:", newBook.clusterId);
+
+      const res = await fetch(`${API_URL}/books`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
@@ -35,98 +41,107 @@ export const useBookStore = create((set) => ({
       });
 
       const data = await res.json();
-      
+
       if (!data.success) {
-        return { success: false, message: data.message || 'Failed to create book' };
+        return {
+          success: false,
+          message: data.message || "Failed to create book",
+        };
       }
 
       set((state) => ({ books: [...state.books, data.data] }));
       return { success: true, message: "Book created successfully" };
     } catch (error) {
-      console.error('Error creating book:', error);
-      return { success: false, message: 'An error occurred while creating the book' };
+      console.error("Error creating book:", error);
+      return {
+        success: false,
+        message: "An error occurred while creating the book",
+      };
     }
   },
 
-  // Fetch books
-  fetchBook: async () => {
+  fetchBook: async (clusterId) => {
     try {
-      set({ booksLoading: true, booksError: null });
-      
-      const res = await fetch("/api/books");
+      set({ booksLoading: true, booksError: null, clusterId });
+
+      const res = await fetch(`${API_URL}/books/cluster/${clusterId}`);
       if (!res.ok) {
         throw new Error(`API responded with status ${res.status}`);
       }
-      
+
       const data = await res.json();
-      
+      console.log("Fetched books data:", data);
+
       if (!data.success) {
-        throw new Error(data.message || 'Failed to fetch books');
+        throw new Error(data.message || "Failed to fetch books");
       }
 
-      set({ 
-        books: data.data,
-        booksLoading: false 
-      });
+      set({ books: data.data || [], booksLoading: false });
+      return data.data;
     } catch (error) {
-      console.error('Error fetching books:', error);
-      set({ 
+      console.error("Error fetching books:", error.message || error);
+      set({
         booksError: error.message,
         booksLoading: false,
-        books: []
+        books: [],
       });
+      return [];
     }
   },
 
   // Delete book
   deleteBook: async (pid) => {
     try {
-      const res = await fetch(`/api/books/${pid}`, {
+      const res = await fetch(`${API_URL}/books/${pid}`, {
         method: "DELETE",
       });
-      
-      const data = await res.json();
-      if (!data.success) {
-        return { success: false, message: data.message };
-      }
 
-      set((state) => ({ 
-        books: state.books.filter((book) => book._id !== pid) 
-      }));
-      
-      return { success: true, message: data.message };
-    } catch (error) {
-      console.error('Error deleting book:', error);
-      return { success: false, message: 'An error occurred while deleting the book' };
-    }
-  },
-
-  // Update book
-  updateBook: async (pid, updatedBook) => {
-    try {
-      const res = await fetch(`/api/books/${pid}`, {
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(updatedBook),
-      });
-      
       const data = await res.json();
       if (!data.success) {
         return { success: false, message: data.message };
       }
 
       set((state) => ({
-        books: state.books.map((book) => 
-          book._id === pid ? data.data : book
-        ),
+        books: state.books.filter((book) => book._id !== pid),
       }));
 
       return { success: true, message: data.message };
     } catch (error) {
-      console.error('Error updating book:', error);
-      return { success: false, message: 'An error occurred while updating the book' };
+      console.error("Error deleting book:", error);
+      return {
+        success: false,
+        message: "An error occurred while deleting the book",
+      };
+    }
+  },
+
+  // Update book
+  updateBook: async (pid, updatedBook) => {
+    try {
+      const res = await fetch(`${API_URL}/books/${pid}`, {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(updatedBook),
+      });
+
+      const data = await res.json();
+      if (!data.success) {
+        return { success: false, message: data.message };
+      }
+
+      set((state) => ({
+        books: state.books.map((book) => (book._id === pid ? data.data : book)),
+      }));
+
+      return { success: true, message: data.message };
+    } catch (error) {
+      console.error("Error updating book:", error);
+      return {
+        success: false,
+        message: "An error occurred while updating the book",
+      };
     }
   },
 
@@ -134,13 +149,13 @@ export const useBookStore = create((set) => ({
   fetchPopularBooks: async () => {
     try {
       set({ popularBooksLoading: true, popularBooksError: null });
-      
-      const baseUrl = import.meta.env.VITE_RAPIDAPI_URL?.replace(/\/+$/, ''); 
+
+      const baseUrl = import.meta.env.VITE_RAPIDAPI_URL?.replace(/\/+$/, "");
       const apiKey = import.meta.env.VITE_RAPIDAPI_KEY;
       const host = import.meta.env.VITE_RAPIDAPI_HOST;
-      
+
       if (!baseUrl || !apiKey || !host) {
-        throw new Error('Missing required API configuration');
+        throw new Error("Missing required API configuration");
       }
 
       const date = new Date();
@@ -151,9 +166,9 @@ export const useBookStore = create((set) => ({
 
       const res = await fetch(url, {
         headers: {
-          'x-rapidapi-key': apiKey,
-          'x-rapidapi-host': host
-        }
+          "x-rapidapi-key": apiKey,
+          "x-rapidapi-host": host,
+        },
       });
 
       if (!res.ok) {
@@ -161,22 +176,22 @@ export const useBookStore = create((set) => ({
       }
 
       const data = await res.json();
-      
+
       if (!Array.isArray(data)) {
-        throw new Error('Unexpected API response format');
+        throw new Error("Unexpected API response format");
       }
-      
-      set({ 
+
+      set({
         popularBooks: data,
-        popularBooksLoading: false 
+        popularBooksLoading: false,
       });
     } catch (error) {
-      console.error('Error fetching popular books:', error);
-      set({ 
+      console.error("Error fetching popular books:", error);
+      set({
         popularBooksError: error.message,
         popularBooksLoading: false,
-        popularBooks: []
+        popularBooks: [],
       });
     }
-  }
+  },
 }));
